+++ src/LinkCollection.php
@@ -7,9 +7,11 @@
 
 namespace Zend\Expressive\Hal;
 
-use Psr\Link\LinkInterface;
 use Psr\Link\EvolvableLinkProviderInterface;
-use Traversable;
+use Psr\Link\LinkInterface;
+
+use function array_filter;
+use function in_array;
 
 /**
  * Properties and methods suitable for an